+++ mne/fiff/__init__.py
@@ -13,5 +13,5 @@ from .pick import pick_types, pick_channels, pick_types_evoked, \
                   pick_types_forward
 
 from .compensator import get_current_comp
-from .proj import compute_spatial_vectors
+from .proj import compute_spatial_vectors, proj_equal
 from .cov import read_cov, write_cov